+++ portal/tests/test_class.py
@@ -46,8 +46,6 @@ from utils.student import create_school_student_directly
 from utils.messages import (
     is_class_created_message_showing,
     is_class_nonempty_message_showing,
-    is_dialog_showing,
-    confirm_dialog,
 )
